+++ packages/api-models/CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   `ResolveObjectsParameters` and `isResolveObjectsParameters`.
+
 ### Changed
+
+-   Renamed `ResolveParameters` to `ResolveObjectParameters` and removed `uuid` property.
+-   Renamed `isResolveParameters` to `isResolveObjectParameters`.
 
 ### Deprecated